+++ Assignment2b/main.cpp
@@ -6,6 +6,6 @@ int main(){
     std::string input;
     std::cout << "Write your expression: ";
     std::getline(std::cin, input);
-    int result = calculator(input);
+    double result = calculator(input);
     std::cout << "Result = " << result << std::endl;
 }
